move main.rs contents into tests/prime500.rs
